yolo-service/requirements.txt
Flask>=3.0.0
ultralytics>=8.1.0
Pillow>=10.0.0
torch>=2.0.0
numpy>=1.24.0
